Update the shell files to check if each file exists

+++ git/init.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
-rm $HOME/.gitconfig
+if [ -f $HOME/.gitconfig ]; then
+  rm $HOME/.gitconfig
+fi
 ln -s $HOME/dotfiles/git/.gitconfig $HOME/.gitconfig
 
-rm $HOME/.gitignore_global
+if [ -f $HOME/.gitignore_global ]; then
+  rm $HOME/.gitignore_global
+fi
 ln -s $HOME/dotfiles/git/.gitignore_global $HOME/.gitignore_global
